+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-04-17 15:29:44
-// Commit: chore(forms): update dependencies and cleanup unused imports in forms
+// Updated at: 2026-04-17 19:16:05
+// Commit: fix(forms): fix memory leak by cancelling forms subscription
 
 class formsComponentState {
-  final String commitRef = '484';
-  final String lastMessage = 'chore(forms): update dependencies and cleanup unused imports in forms';
-  final DateTime timestamp = DateTime.parse('2026-04-17T15:29:44+05:00');
+  final String commitRef = '492';
+  final String lastMessage = 'fix(forms): fix memory leak by cancelling forms subscription';
+  final DateTime timestamp = DateTime.parse('2026-04-17T19:16:05+05:00');
 }